backend-core/requirements.txt
django~=4.1.7
djangorestframework-simplejwt==5.2.0
django-simple-history==3.2.0
bpython
djangorestframework
django-filter
psycopg2-binary
django-cors-headers
drf-writable-nested
django_elasticsearch_dsl
elasticsearch_dsl
shortuuid
Werkzeug
python-decouple==3.7
